src/main/java/com/study/realworld/domain/article/application/ArticleQueryService.java
package com.study.realworld.domain.article.application;

import com.study.realworld.domain.article.domain.persist.Article;
import com.study.realworld.domain.article.domain.persist.ArticleRepository;
import com.study.realworld.domain.article.domain.vo.ArticleSlug;
import com.study.realworld.domain.article.dto.ArticleSave;
import com.study.realworld.domain.user.application.UserQueryService;
import com.study.realworld.domain.user.domain.persist.User;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@RequiredArgsConstructor
@Transactional(readOnly = true)
@Service
public class ArticleQueryService {

    private final ArticleRepository articleRepository;

    public Article findByArticleSlug(final ArticleSlug articleSlug) {
        return articleRepository.findByArticleSlug(articleSlug)
                .orElseThrow(IllegalArgumentException::new);
    }
}
